+++ src/app/actions/auth.ts
@@ -1,0 +1,7 @@
+"use server";
+
+import { getSession } from "../lib/session";
+
+export async function refreshSession() {
+  return await getSession();
+}
