+++ src/app/api/login/route.ts
@@ -59,7 +59,7 @@ export async function POST(request: Request) {
       );
     }
 
-    // Create a session with user data
+    // Create session data
     const session = {
       userId: user.id,
       user: {
@@ -67,18 +67,12 @@ export async function POST(request: Request) {
         email: user.email,
         name: user.name,
         role: user.userRole,
-        image: user.image,
+        image: user.image || "/default-avatar.webp", // Ensure we always have an image URL
       },
     };
 
     // Encrypt session and set cookie
-    const encryptedSession = await encrypt({
-      ...session,
-      user: {
-        ...session.user,
-        image: user.image || undefined,
-      },
-    });
+    const encryptedSession = await encrypt(session);
     (await cookies()).set("session", encryptedSession, {
       httpOnly: true,
       secure: process.env.NODE_ENV === "production",
