+++ src/app/login/page.tsx
@@ -1,71 +1,25 @@
-"use client";
+import { getSession } from "../lib/session";
+import { redirect } from "next/navigation";
+import LoginForm from "./LoginForm";
 
-import { FormEvent, useState, useEffect } from "react";
-import { useRouter, useSearchParams } from "next/navigation";
-import Link from "next/link";
-import Image from "next/image";
-import MagicLinkForm from "./magic-link-form";
-
-export default function LoginPage() {
-  const router = useRouter();
-  const searchParams = useSearchParams();
-  const [error, setError] = useState<string | null>(null);
-  const [success, setSuccess] = useState<string | null>(null);
-  const [isLoading, setIsLoading] = useState(false);
-
-  // Check if user just registered or password reset was successful
-  useEffect(() => {
-    if (searchParams.get("registered") === "true") {
-      setSuccess("Account created successfully! Please sign in.");
-    }
-    
-    if (searchParams.get("reset") === "success") {
-      setSuccess("Password reset successful! Please sign in.");
-    }
-  }, [searchParams]);
-
-  async function handleSubmit(event: FormEvent<HTMLFormElement>) {
-    event.preventDefault();
-    setIsLoading(true);
-    setError(null);
-    setSuccess(null);
-
-    const formData = new FormData(event.currentTarget);
-    const email = formData.get("email") as string;
-    const password = formData.get("password") as string;
-
-    try {
-      const response = await fetch("/api/login", {
-        method: "POST",
-        headers: { "Content-Type": "application/json" },
-        body: JSON.stringify({ email, password }),
-      });
-
-      if (response.ok) {
-        // Redirect to dashboard on successful login
-        router.push("/dashboard");
-        router.refresh();
-      } else {
-        const data = await response.json();
-        setError(data.error || "Login failed");
-      }
-    } catch (err) {
-      console.error("Login error:", err);
-      setError("An error occurred. Please try again.");
-    } finally {
-      setIsLoading(false);
-    }
+export default async function LoginPage({
+  searchParams,
+}: {
+  searchParams: { registered?: string; reset?: string };
+}) {
+  // Check if user is already logged in
+  const session = await getSession();
+  if (session) {
+    redirect("/dashboard");
   }
 
   return (
     <div className="flex min-h-screen flex-col items-center justify-center p-4 relative">
       <div className="absolute inset-0 z-0">
-        <Image
+        <img
           src="/chemex-brewing-landing.png"
           alt="Coffee brewing background"
-          fill
-          className="object-cover opacity-80"
-          priority
+          className="object-cover opacity-80 w-full h-full"
         />
       </div>
       <div className="w-full max-w-md space-y-8 rounded-lg border p-6 shadow-md bg-base-100/90 relative z-10">
@@ -73,80 +27,21 @@ export default function LoginPage() {
           <h1 className="text-2xl font-bold">Sign in to your account</h1>
           <p className="mt-2 text-sm text-gray-600 coffee:text-gray-400">
             Or{" "}
-            <Link href="/register" className="text-blue-600 hover:underline">
+            <a href="/register" className="text-blue-600 hover:underline">
               create a new account
-            </Link>
+            </a>
           </p>
         </div>
 
-        {error && (
-          <div className="alert alert-error">
-            {error}
-          </div>
-        )}
-
-        {success && (
-          <div className="alert alert-success">
-            {success}
-          </div>
-        )}
-
-        <div className="flex justify-between items-center mt-4">
-          <Link href="/forgot-password" className="text-sm text-primary hover:underline">
-            Forgot password?
-          </Link>
-          {/* Remove the magic link here since we have the collapsible form below */}
-        </div>
-
-        <form className="mt-8 space-y-6" onSubmit={handleSubmit}>
-          <div className="space-y-4">
-            <div>
-              <label htmlFor="email" className="block text-sm font-medium">
-                Email address
-              </label>
-              <input
-                id="email"
-                name="email"
-                type="email"
-                autoComplete="email"
-                required
-                className="input input-bordered w-full"
-                placeholder="you@example.com"
-              />
-            </div>
-            <div>
-              <label htmlFor="password" className="block text-sm font-medium">
-                Password
-              </label>
-              <input
-                id="password"
-                name="password"
-                type="password"
-                autoComplete="current-password"
-                required
-                className="input input-bordered w-full"
-                placeholder="••••••••"
-              />
-            </div>
-            <div className="flex justify-end">
-              <Link href="/forgot-password" className="text-sm text-primary hover:underline">
-                Forgot password?
-              </Link>
-            </div>
-          </div>
-
-          <div>
-            <button
-              type="submit"
-              disabled={isLoading}
-              className="btn btn-primary w-full"
-            >
-              {isLoading ? "Signing in..." : "Sign in"}
-            </button>
-          </div>
-        </form>
-
-        <MagicLinkForm />
+        <LoginForm
+          successMessage={
+            (await searchParams).registered === "true"
+              ? "Account created successfully! Please sign in."
+              : (await searchParams).reset === "success"
+                ? "Password reset successful! Please sign in."
+                : null
+          }
+        />
       </div>
     </div>
   );
