+++ src/app/register/page.tsx
@@ -1,117 +1,21 @@
-"use client";
+import { getSession } from "../lib/session";
+import { redirect } from "next/navigation";
+import RegisterForm from "@/app/register/RegisterForm";
 
-import { FormEvent, useState } from "react";
-import { useRouter } from "next/navigation";
-import Link from "next/link";
-import { z } from "zod";
-import Image from "next/image";
-
-// Client-side validation schema
-const registerFormSchema = z
-  .object({
-    name: z
-      .string()
-      .min(2, { message: "Name must be at least 2 characters" })
-      .max(50, { message: "Name cannot be longer than 50 characters" }),
-    email: z.string().email({ message: "Valid email is required" }),
-    password: z
-      .string()
-      .min(8, { message: "Password must be at least 8 characters" })
-      .regex(/[a-z]/, {
-        message: "Password must contain at least one lowercase letter",
-      })
-      .regex(/[A-Z]/, {
-        message: "Password must contain at least one uppercase letter",
-      })
-      .regex(/[0-9]/, { message: "Password must contain at least one number" }),
-    confirmPassword: z.string(),
-  })
-  .refine((data) => data.password === data.confirmPassword, {
-    message: "Passwords do not match",
-    path: ["confirmPassword"],
-  });
-
-export default function RegisterPage() {
-  const router = useRouter();
-  const [errors, setErrors] = useState<{ [key: string]: string }>({});
-  const [isLoading, setIsLoading] = useState(false);
-  const [generalError, setGeneralError] = useState<string | null>(null);
-
-  async function handleSubmit(event: FormEvent<HTMLFormElement>) {
-    event.preventDefault();
-    setIsLoading(true);
-    setErrors({});
-    setGeneralError(null);
-
-    const formData = new FormData(event.currentTarget);
-    const formValues = {
-      name: formData.get("name") as string,
-      email: formData.get("email") as string,
-      password: formData.get("password") as string,
-      confirmPassword: formData.get("confirmPassword") as string,
-    };
-
-    // Client-side validation using Zod
-    const validationResult = registerFormSchema.safeParse(formValues);
-
-    if (!validationResult.success) {
-      const formattedErrors: { [key: string]: string } = {};
-      validationResult.error.errors.forEach((error) => {
-        formattedErrors[error.path.join(".")] = error.message;
-      });
-
-      setErrors(formattedErrors);
-      setIsLoading(false);
-      return;
-    }
-
-    try {
-      const response = await fetch("/api/register", {
-        method: "POST",
-        headers: { "Content-Type": "application/json" },
-        body: JSON.stringify({
-          name: formValues.name,
-          email: formValues.email,
-          password: formValues.password,
-        }),
-      });
-
-      const data = await response.json();
-
-      if (response.ok) {
-        // Registration successful, redirect to login
-        router.push("/login?registered=true");
-      } else {
-        // Handle server validation errors
-        if (data.errors) {
-          const serverErrors: { [key: string]: string } = {};
-          Object.entries(data.errors).forEach(([key, messages]) => {
-            serverErrors[key] = Array.isArray(messages)
-              ? messages[0]
-              : messages;
-          });
-          setErrors(serverErrors);
-        } else {
-          setGeneralError(data.error || "Registration failed");
-        }
-      }
-    } catch (err) {
-      console.error("Registration error:", err);
-      setGeneralError("An error occurred. Please try again.");
-    } finally {
-      setIsLoading(false);
-    }
+export default async function RegisterPage() {
+  // Check if user is already logged in
+  const session = await getSession();
+  if (session) {
+    redirect("/dashboard");
   }
 
   return (
     <div className="flex min-h-screen flex-col items-center justify-center p-4 relative">
       <div className="absolute inset-0 z-0">
-        <Image
+        <img
           src="/chemex-brewing-landing.png"
           alt="Coffee brewing background"
-          fill
-          className="object-cover opacity-80"
-          priority
+          className="object-cover opacity-80 w-full h-full"
         />
       </div>
       <div className="w-full max-w-md space-y-8 rounded-lg border p-6 shadow-md bg-base-100/90 relative z-10">
@@ -119,111 +23,13 @@ export default function RegisterPage() {
           <h1 className="text-2xl font-bold">Create an account</h1>
           <p className="mt-2 text-sm text-gray-600 coffee:text-gray-400">
             Already have an account?{" "}
-            <Link href="/login" className="text-blue-600 hover:underline">
+            <a href="/login" className="text-blue-600 hover:underline">
               Sign in
-            </Link>
+            </a>
           </p>
         </div>
 
-        {generalError && (
-          <div className="alert alert-error">
-            {generalError}
-          </div>
-        )}
-
-        <form className="mt-8 space-y-6" onSubmit={handleSubmit}>
-          <div className="space-y-4">
-            <div>
-              <label htmlFor="name" className="block text-sm font-medium">
-                Full Name
-              </label>
-              <input
-                id="name"
-                name="name"
-                type="text"
-                autoComplete="name"
-                required
-                className="input input-bordered w-full"
-                placeholder="John Doe"
-              />
-              {errors.name && (
-                <p className="mt-1 text-sm text-red-600">{errors.name}</p>
-              )}
-            </div>
-
-            <div>
-              <label htmlFor="email" className="block text-sm font-medium">
-                Email address
-              </label>
-              <input
-                id="email"
-                name="email"
-                type="email"
-                autoComplete="email"
-                required
-                className="input input-bordered w-full"
-                placeholder="you@example.com"
-              />
-              {errors.email && (
-                <p className="mt-1 text-sm text-red-600">{errors.email}</p>
-              )}
-            </div>
-
-            <div>
-              <label htmlFor="password" className="block text-sm font-medium">
-                Password
-              </label>
-              <input
-                id="password"
-                name="password"
-                type="password"
-                autoComplete="new-password"
-                required
-                className="input input-bordered w-full"
-                placeholder="••••••••"
-              />
-              {errors.password && (
-                <p className="mt-1 text-sm text-red-600">{errors.password}</p>
-              )}
-              <p className="mt-1 text-xs text-gray-500">
-                Must be at least 8 characters
-              </p>
-            </div>
-
-            <div>
-              <label
-                htmlFor="confirmPassword"
-                className="block text-sm font-medium"
-              >
-                Confirm Password
-              </label>
-              <input
-                id="confirmPassword"
-                name="confirmPassword"
-                type="password"
-                autoComplete="new-password"
-                required
-                className="input input-bordered w-full"
-                placeholder="••••••••"
-              />
-              {errors.confirmPassword && (
-                <p className="mt-1 text-sm text-red-600">
-                  {errors.confirmPassword}
-                </p>
-              )}
-            </div>
-          </div>
-
-          <div>
-            <button
-              type="submit"
-              disabled={isLoading}
-              className="btn btn-primary w-full"
-            >
-              {isLoading ? "Creating account..." : "Create account"}
-            </button>
-          </div>
-        </form>
+        <RegisterForm />
       </div>
     </div>
   );
